functions/supabase_archiver/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
supabase==2.*
